tests/wpt/web-platform-tests/app-history/navigate/navigate-same-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="d"></div>
<script>
promise_test(async () => {
  let onnavigate_called = false;
  appHistory.onnavigate = () => onnavigate_called = true;
  await appHistory.navigate("#d");
  assert_equals(location.hash, "#d");
  assert_true(onnavigate_called);
  assert_equals(document.querySelector(":target"), d);
}, "navigate() navigates same-document and fires onnavigate (async)");

test(() => {
  let onnavigate_called = false;
  appHistory.onnavigate = () => onnavigate_called = true;
  appHistory.navigate("#d");
  assert_equals(location.hash, "#d");
  assert_true(onnavigate_called);
  assert_equals(document.querySelector(":target"), d);
}, "navigate() navigates same-document and fires onnavigate (sync)");
</script>
